Extract creator-conversation MCP tools out of mcp-server.ts
Continues Wave C (batch 3, mcp-server.ts 5164 -> 4817 lines). Moves
read_inbox/get_transcript/ack_inbox into mcp-inbox-tools.ts as
createInboxTools({resolveAuth, injectChannel, writePiggyback}).

Promotes five pure, module-level constructs into mcp-tool-support.ts
that this cluster shares with ~20 other tool sites still in
mcp-server.ts (pendingMessagesFromChannel, ChannelControlBody,
channelControlFields, CREATOR_TEXT_SAFETY, WARNINGS_PROP,
REPLY_CONTROL) — a verbatim move with no external call-site edits,
since those sites just pick up the same identifiers via import.
writePiggyback stays in mcp-server.ts and is injected as a
dependency, since it genuinely closes over injectChannel.
